backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# AI Assistant dependencies
langchain-groq==0.1.5
langchain-core==0.1.0
langchain-community==0.0.38
pillow==10.0.0
pypdf2==3.0.1
aiofiles==23.2.1